Restartable sequences: clear rseq_cs even if non-nested
We can speed up the user-space fast path by removing the final
"rseq_cs=NULL" assignment at the end of the critical section, since we
check for both upper and lower bounds anyway in the fixup code.

Take care of setting rseq_cs back to NULL even if we are not nested over
an actual rseq critical section.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>