Restartable sequences: don't clear rseq_cs after c.s.
We can remove this rseq_cs=NULL clearing step at the end of the critical
section, since the kernel checks for lower and upper address ranges when
it preempts or delivers a signals. The kernel will either take care of
clearing it at the next preemption or signal delivery, or the next
critical section will either change its value, or set it to the same
value.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>