Restartable sequences: tests: introduce simple rseq start/finish
Introduce rseq_start/rseq_finish that do not require the rseq lock. The
new rseq_start_rlock and rseq_finish_rlock deal with it.

This is useful for use-cases that do not require locking to handle
fallback. For instance, using split counter techniques for fallback when
possible, or using reference counting to do a lock-free fallback.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>