Restartable sequences selftests: use initial-exec TLS model, non-weak symbol
The initial-exec model is signal-safe, whereas the global-dynamic model
is not.

Remove the "weak" symbol attribute from the __rseq_abi in rseq.c. The
rseq.so library will have ownership of that symbol, and there is not
reason for an application or user library to try to define that
symbol.

The expected use is to link against libreq.so, which owns and provide that
symbol.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>