Restartable sequences selftests: bail out compiling param_test with gcc <= 4.8 on x86-32
gcc 4.8 and prior fails at compiling the parametrized test due to
register constraints issues in O2 on x86-32.

gcc-4.9 and better are fine with it. Given that this extra register
pressure is caused by the extra scratch space we keep for delay loop in
the parametrized test (this is only for a delay-insertion build), simply
tell users to upgrade their compiler.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>