cpu_mutex: Introduce CPU mutual exclusion system call
There is an important use-case which is not possible with the
"rseq" (Restartable Sequences) system call, which was left as
future work.

That use-case is to modify user-space per-cpu data structures
belonging to specific CPUs which may be brought offline and
online again by CPU hotplug. This can be used by memory
allocators to migrate free memory pools when CPUs are brought
offline, or by ring buffer consumers to target specific per-CPU
buffers, even when CPUs are brought offline.

A few rather complex prior attempts were made to solve this.
Those were based on in-kernel interpreters (cpu_opv, do_on_cpu).
That complexity was generally frowned upon, even by their author.

This patch fulfills this use-case in a refreshingly simple way:
it introduces a "cpu_mutex" system call, which allows user-space
threads to ask to run with mutual exclusion with respect to
user-space execution on a specific CPU, and then clear this state.

"But this can already be done with sched_setaffinity", some
would rightfully reply. However, sched_setaffinity is not safe
with respect to concurrent cpu hotplug and cpusets changes.
This new "cpu_mutex" concept addresses those limitations.

The key idea behind "cpu mutex" is to let a worker thread eat
CPU time on the target CPU while the user-space thread is provided
"mutually exclusive" access to user-space per-cpu data of that CPU,
even though it is running on another CPU. If any of the thread or the
worker thread are preempted, both get preempted.

For instance, this allows implementing this userspace library API
for incrementing a per-cpu counter for a specific cpu number
received as parameter:

static inline __attribute__((always_inline))
int percpu_addv(intptr_t *v, intptr_t count, int cpu)
{
        int ret;

        ret = rseq_addv(v, count, cpu);
check:
        if (rseq_unlikely(ret)) {
                pin_on_cpu_set(cpu);
                ret = rseq_addv(v, count, percpu_current_cpu());
                pin_on_cpu_clear();
                goto check;
        }
        return 0;
}

Not-signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>