rseq: Introduce RSEQ_FLAG_REGISTER
Introduce RSEQ_FLAG_REGISTER with the same behavior as the flag value
"0". The main advantage of introducing this flag as a non-zero (1 << 1)
value is that it can be combined with other flags to register and check
for features with a single system call. Considering that this system
call needs to be performed for each new thread in glibc, minimize the
amount of overhead required.

This is needed for introducing a new RSEQ_FLAG_RELIABLE_CPU_ID flag in a
later change.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Florian Weimer <fw@deneb.enyo.de>
Cc: "Paul E. McKenney" <paulmck@linux.ibm.com>
Cc: Boqun Feng <boqun.feng@gmail.com>
Cc: "H . Peter Anvin" <hpa@zytor.com>
Cc: Paul Turner <pjt@google.com>
Cc: Dmitry Vyukov <dvyukov@google.com>
Cc: Neel Natu <neelnatu@google.com>
Cc: linux-api@vger.kernel.org