rseq: Introduce RSEQ_FLAG_RELIABLE_CPU_ID
commit 93b585c08d16 ("Fix: sched: unreliable rseq cpu_id for new tasks")
addresses an issue with cpu_id field of newly created processes. Expose
a flag which can be used by user-space to query whether the kernel
implements this fix.

For use-cases where using rseq with the wrong cpu_id would lead to
corruption of user-space per-cpu data, it is recommended that user-space
detects availability of this fix by calling rseq with the
RSEQ_FLAG_RELIABLE_CPU_ID flag.

When the rseq TLS is shared across libraries, for instance if the rseq
TLS ABI is registered by the libc, applications and libraries should
call rseq with this flag to validate that the kernel implements a
reliable cpu_id field.

It can also be OR'd with RSEQ_FLAG_REGISTER, so an application managing
the entire ownership of its TLS registration can register the rseq TLS
and query for reliable cpu_id at the same time.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Florian Weimer <fw@deneb.enyo.de>
Cc: "Paul E. McKenney" <paulmck@linux.ibm.com>
Cc: Boqun Feng <boqun.feng@gmail.com>
Cc: "H . Peter Anvin" <hpa@zytor.com>
Cc: Paul Turner <pjt@google.com>
Cc: Dmitry Vyukov <dvyukov@google.com>
Cc: Neel Natu <neelnatu@google.com>
Cc: linux-api@vger.kernel.org